# swagger-output.json
{
    "openapi": "3.0.0",
    "info": {
        "version": "1.0.0",
        "title": "Nexalog"
    },
    "servers": [
        {
            "url": "http://localhost:4000"
        }
    ],
...
    "paths": {
...
"/api/chat/conversations/{conversationId}/messages": {
  "get": {
    "tags": ["Czat prywatny"],
    "summary": "Pobierz ostatnie wiadomości z konwersacji",
    "security": [{ "bearerAuth": [] }],
    "parameters": [
      { "name": "conversationId", "in": "path", "required": true, "schema": { "type": "integer" } },
      { "name": "limit", "in": "query", "required": false, "schema": { "type": "integer", "default": 30, "maximum": 100 } },
      { "name": "beforeId", "in": "query", "required": false, "schema": { "type": "integer" }, "description": "Kursor – pobierz wiadomości starsze niż to ID" }
    ],
    "responses": {
      "200": {
        "description": "OK",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "messages": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "ID_Message": { "type": "integer" },
                      "ID_Conversation": { "type": "integer" },
                      "ID_Sender": { "type": "integer" },
                      "content": { "type": "string" },
                      "edited": { "type": "boolean" },
                      "deleted": { "type": "boolean" },
                      "createdAt": { "type": "string", "format": "date-time" }
                    }
                  }
                },
                "nextBeforeId": { "type": "integer", "nullable": true }
              }
            }
          }
        }
      },
      "403": { "description": "Brak dostępu" },
      "404": { "description": "Konwersacja nie istnieje" }
    }
  },
  "post": {
    "tags": ["Czat prywatny"],
    "summary": "Wyślij wiadomość do konwersacji",
    "security": [{ "bearerAuth": [] }],
    "parameters": [
      { "name": "conversationId", "in": "path", "required": true, "schema": { "type": "integer" } }
    ],
    "requestBody": {
      "required": true,
      "content": {
        "application/json": {
          "schema": {
            "type": "object",
            "properties": {
              "content": { "type": "string", "example": "Cześć!" }
            },
            "required": ["content"]
          }
        }
      }
    },
    "responses": {
      "201": {
        "description": "Wysłano",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "message": { "type": "string", "example": "Wiadomość wysłana." },
                "data": {
                  "type": "object",
                  "properties": {
                    "ID_Message": { "type": "integer" },
                    "ID_Conversation": { "type": "integer" },
                    "ID_Sender": { "type": "integer" },
                    "content": { "type": "string" },
                    "createdAt": { "type": "string", "format": "date-time" }
                  }
                }
              }
            }
          }
        }
      },
      "400": { "description": "Błąd walidacji (np. pusta treść)" },
      "403": { "description": "Brak dostępu" },
      "404": { "description": "Konwersacja nie istnieje" }
    }
  }
},
"/api/chat/conversations/{conversationId}/messages/{messageId}": {
  "put": {
    "tags": ["Czat prywatny"],
    "summary": "Edytuj własną wiadomość",
    "description": "Pozwala zmienić treść **własnej** wiadomości w danej konwersacji. Zwraca zaktualizowany obiekt wiadomości.",
    "security": [{ "bearerAuth": [] }],
    "parameters": [
      {
        "name": "conversationId",
        "in": "path",
        "required": true,
        "schema": { "type": "integer", "example": 1 }
      },
      {
        "name": "messageId",
        "in": "path",
        "required": true,
        "schema": { "type": "integer", "example": 123 }
      }
    ],
    "requestBody": {
      "required": true,
      "content": {
        "application/json": {
          "schema": {
            "type": "object",
            "properties": {
              "content": { "type": "string", "maxLength": 5000, "example": "Poprawiona treść wiadomości" }
            },
            "required": ["content"]
          }
        }
      }
    },
    "responses": {
      "200": {
        "description": "Wiadomość zaktualizowana",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "message": { "type": "string", "example": "Wiadomość zaktualizowana" },
                "data": {
                  "type": "object",
                  "properties": {
                    "ID_Message": { "type": "integer", "example": 123 },
                    "ID_Conversation": { "type": "integer", "example": 1 },
                    "ID_Sender": { "type": "integer", "example": 7 },
                    "content": { "type": "string", "example": "Poprawiona treść wiadomości" },
                    "edited": { "type": "boolean", "example": true },
                    "deleted": { "type": "boolean", "example": false },
                    "createdAt": { "type": "string", "format": "date-time" },
                    "updatedAt": { "type": "string", "format": "date-time" }
                  }
                }
              }
            }
          }
        }
      },
      "400": {
        "description": "Błąd walidacji (np. pusta treść / zbyt długa / błędne ID)",
        "content": {
          "application/json": {
            "schema": { "type": "object", "properties": { "message": { "type": "string", "example": "Treść wiadomości jest wymagana" } } }
          }
        }
      },
      "403": {
        "description": "Brak uprawnień (nie jesteś autorem albo nie należysz do konwersacji)",
        "content": {
          "application/json": {
            "schema": { "type": "object", "properties": { "message": { "type": "string", "example": "Możesz modyfikować tylko własne wiadomości." } } }
          }
        }
      },
      "404": {
        "description": "Nie znaleziono konwersacji lub wiadomości",
        "content": {
          "application/json": {
            "schema": { "type": "object", "properties": { "message": { "type": "string", "example": "Wiadomość nie istnieje." } } }
          }
        }
      },
      "500": {
        "description": "Błąd serwera",
        "content": {
          "application/json": {
            "schema": { "type": "object", "properties": { "message": { "type": "string", "example": "Wewnętrzny błąd serwera" } } }
          }
        }
      }
    }
  },
  "delete": {
    "tags": ["Czat prywatny"],
    "summary": "Usuń własną wiadomość (soft delete)",
    "description": "Oznacza wiadomość jako usuniętą (flaga `deleted=true`, treść pusta). Idempotentne – ponowny DELETE zwróci 204.",
    "security": [{ "bearerAuth": [] }],
    "parameters": [
      {
        "name": "conversationId",
        "in": "path",
        "required": true,
        "schema": { "type": "integer", "example": 1 }
      },
      {
        "name": "messageId",
        "in": "path",
        "required": true,
        "schema": { "type": "integer", "example": 123 }
      }
    ],
    "responses": {
      "204": { "description": "Wiadomość usunięta (soft delete)" },
      "403": {
        "description": "Brak uprawnień",
        "content": {
          "application/json": {
            "schema": { "type": "object", "properties": { "message": { "type": "string", "example": "Możesz modyfikować tylko własne wiadomości." } } }
          }
        }
      },
      "404": {
        "description": "Nie znaleziono",
        "content": {
          "application/json": {
            "schema": { "type": "object", "properties": { "message": { "type": "string", "example": "Wiadomość nie istnieje." } } }
          }
        }
      },
      "500": {
        "description": "Błąd serwera",
        "content": {
          "application/json": {
            "schema": { "type": "object", "properties": { "message": { "type": "string", "example": "Wewnętrzny błąd serwera" } } }
          }
        }
      }
    }
  }
},
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        },
...
    }
}